+++ src/types/SDKResponse.ts
@@ -8,19 +8,37 @@ import { FetchError } from "../library/FetchError";
 export type SDKResponse<T> =
 	| {
 			/**
-			 * The data returned from the API call. The type is set in the generic argument in the {@link SDKResponse}.
-			 */
-			data: T;
-			/**
 			 * A boolean set to false to indicate the API call was successful.
 			 */
 			isError: false;
+			/**
+			 * An object with properties to help identify requests in logging.
+			 */
+			tracing: {
+				/**
+				 * The request ID used to identify requests in logs and for distributed tracing.
+				 */
+				frontasticRequestId?: string;
+			};
+			/**
+			 * The data returned from the API call. The type is set in the generic argument in the {@link SDKResponse}.
+			 */
+			data: T;
 	  }
 	| {
 			/**
 			 * A boolean set to true to indicate an error occurred during the API call.
 			 */
 			isError: true;
+			/**
+			 * An object with properties to help identify requests in logging.
+			 */
+			tracing: {
+				/**
+				 * The request ID used to identify requests in logs and for distributed tracing. Only provided when the request reaches the extension runner.
+				 */
+				frontasticRequestId?: string;
+			};
 			/**
 			 * The error which occurred during the API call.
 			 */