+++ src/types/api/page/PageResponse.ts
@@ -2,10 +2,10 @@ import { Page, PageFolder } from "@frontastic/extension-types";
 import { PageViewData } from "./PageViewData";
 
 //TODO: update @frontastic/extension-types to include PageResponse
-type PageResponse = {
+interface PageResponse {
 	page: Page;
 	pageFolder: PageFolder;
 	data: PageViewData;
-};
+}
 
 export { PageResponse };